add loot function

+++ components/enemy.js
@@ -5,39 +5,34 @@ export class Enemy {
         this.x = x
         this.y = y
         this.level = level
-        //this.id = i
+
         this.direction = "Right"
         this.enemySize = 40
-        //console.log(this.enemySize)
         this.speed = 2
         this.detect = true
         this.lastposition = ""
         this.mustrender = true
         this.targetX = x
         this.targetY = y
-        // I will accesS to The Target div enemys !
         this.Div = null
         this.dead = false
-        // this.creatediv()
-        // her i add cordination frame up left right !!
         this.AnimationCord = Cordination
     }
-    // i will remove enenmy
-    killenemy(cs = true) {
+
+    killEnemy(cs = true) {
         if (this.Div && this.Div.parentNode) {
             this.Div.parentNode.removeChild(this.Div);
-            // this.game.removeEnemy(this);
             if (cs) this.game.state.setScore(100)
             this.dead = true;
         }
-        // Clear references
+
         this.Div = null;
         this.game = null;
         this.level = null;
         this.AnimationCord = null;
     }
 
-    checkiftheriscolision() {
+    checkColision() {
         const blockSize = this.level.block_size;
         const now = performance.now();
         for (let bomb of this.game.map.bombs) {
@@ -47,10 +42,8 @@ export class Enemy {
             const bombY = bomb.yMap * blockSize;
 
             if (this.isColliding(bombX, bombY, blockSize, blockSize)) {
-                // i want to test it and remove the div of enemys where we are now how i will do it exect
-                this.killenemy()
+                this.killEnemy()
                 this.dead = true
-                //this.kill();
                 return;
             }
             for (let dir of bomb.freeBlocks) {
@@ -61,7 +54,7 @@ export class Enemy {
                 if (dir === 3) ex += blockSize;
 
                 if (this.isColliding(ex, ey, blockSize, blockSize)) {
-                    this.killenemy();
+                    this.killEnemy();
                     this.dead = true;
                     return;
                 }
@@ -69,19 +62,14 @@ export class Enemy {
         }
     }
 
-    Canmoveandupdate() {
+    updateRender() {
         if (this.dead) return
        
-        this.checkiftheriscolision()
-        // i get problem of it's get access to player memory while he is null  and it's need to checkk first !
-        if (!this.game || !this.game.player) {
-            return; // Safe exit - don't try to use null objects
-        }
-        //this.checkiftheriscolision()
+        this.checkColision()
+        if (!this.game || !this.game.player)  return
          
-        if (this.game.player.isColliding(this.x, this.y, this.enemySize, this.enemySize)) {
-            this.game.player.kill()
-        }
+        if (this.game.player.isColliding(this.x, this.y, this.enemySize, this.enemySize)) this.game.player.kill()
+
         const directions = {
             Up: { rowset: -1, colset: 0 },
             Down: { rowset: 1, colset: 0 },
@@ -90,18 +78,15 @@ export class Enemy {
         }
 
         const blockSize = this.level.block_size
-        //  her i face i round it becs it's give me exect traget col and row in all Grid !!
         const col = (this.direction === "Left") ? Math.round(this.x / blockSize) : Math.floor(this.x / blockSize);
         const row = (this.direction === "Left") ? Math.round(this.y / blockSize) : Math.floor(this.y / blockSize);
         if (this.detect) {
             const nextRow = row + directions[this.direction].rowset
             const nextCol = col + directions[this.direction].colset
 
-            // check if i can move !!
             if (this.game.map.Canmove(nextRow, nextCol)) {
                 this.targetX = nextCol * blockSize + 12
                 this.targetY = nextRow * blockSize + 12
-                //this.lastposition = this.direction
             } else {
                 this.detect = false
                 this.lastposition = this.direction
@@ -114,34 +99,16 @@ export class Enemy {
                 this.direction = this.randomDirection()
             }
         }
-        // Move pixel by pixel to the the next tiles but not in one shoot !
         if (this.x < this.targetX) this.x += this.speed
         if (this.x > this.targetX) this.x -= this.speed
         if (this.y < this.targetY) this.y += this.speed
         if (this.y > this.targetY) this.y -= this.speed
 
-        // get the value of the target tiles by moving !!
         if (Math.abs(this.x - this.targetX) < this.speed) this.x = this.targetX
         if (Math.abs(this.y - this.targetY) < this.speed) this.y = this.targetY
-        //this.updateVisualPosition()
         this.arzigid()
-        //this.updateVisualPosition()
     }
-    /*
-    MOVE() {
 
-        const scale = this.game.map.currentScale || 1;
-        // IF EXEIST
-        /// <===> so what's i sould do now 
-        console.log("=>>",this.direction , this.AnimationCord[this.direction].y, scale, this.game.map.currentScale)
-        this.Div.style.backgroundPosition = `${this.AnimationCord[this.direction].x} ${this.AnimationCord[this.direction].y}`
-        // <========>|<->|<========>
-        //this.Div.style.backgroundPosition = `${this.AbackgroundPositionnimationCord[this.direction].x}px ${this.AnimationCord[this.direction].y}px`;
-        // Mehdi her i can make it layer who is isolated from other layer for more performance !!
-        //this.Div.style.transform = `translate3d(${this.x}px, ${this.y}px,10px)`;
-        //this.updateVisualPosition()this.game.map.currentScale
-    }
-    */
     arzigid() {
         if (!this.Div) return;
         const scale = this.game.map.currentScale || 1;
@@ -160,21 +127,8 @@ export class Enemy {
         return dirs[Math.floor(Math.random() * dirs.length)]
     }
 
-    // helper to check collision with enemy 
     isColliding(x, y, w, h) {
         return !(this.x + this.enemySize < x || this.x > x + w ||
             this.y + this.enemySize < y || this.y > y + h);
     }
-
-    /// update scale !!
-    /*
-    updateVisualPosition() {
-        //  i dont know why he is not chnage there animation while i try alot
-        if (this.Div) {
-            const scale = this.game.map.currentScale || 1;
-            this.Div.style.transform = `translate(${this.x * scale}px, ${this.y * scale}px)`;
-        }
-    }
-    */
-    // HER I CLEAN DOM !! 
 }
